♻️ 36REEV refactor: split PR metadata helpers
