+++ .agentplane/tasks/202605132058-36REEV/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605132058-36REEV`
+Title: Split PR metadata helpers
+Canonical task record: `.agentplane/tasks/202605132058-36REEV/README.md`
+
+## Summary
+
+Split PR metadata helpers
+
+Refactor the shared branch_pr PR metadata helper into smaller compatibility-preserving modules without changing public imports or behavior.
+
+## Scope
+
+- In scope: Refactor the shared branch_pr PR metadata helper into smaller compatibility-preserving modules without changing public imports or behavior.
+- Out of scope: unrelated refactors not required for "Split PR metadata helpers".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-13T20:59:57.460Z
+- Branch: task/202605132058-36REEV/split-pr-meta
+- Head: 403e10e9a434
+
+```text
+No changes detected.
+```
+
+</details>
