+++ .agentplane/tasks/202605132058-36REEV/pr/meta.json
@@ -1,0 +1,14 @@
+{
+  "base": "main",
+  "branch": "task/202605132058-36REEV/split-pr-meta",
+  "created_at": "2026-05-13T20:59:57.460Z",
+  "head_sha": "403e10e9a434aeafa4ac7624336e6a4fee051293",
+  "last_verified_at": null,
+  "last_verified_sha": null,
+  "schema_version": 1,
+  "task_id": "202605132058-36REEV",
+  "updated_at": "2026-05-13T20:59:57.460Z",
+  "verify": {
+    "status": "skipped"
+  }
+}
