+++ .agentplane/tasks/202605132058-36REEV/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-13T20:59:57.460Z
+
+## Task
+
+- Task: `202605132058-36REEV`
+- Title: Split PR metadata helpers
+- Status: DOING
+- Branch: `task/202605132058-36REEV/split-pr-meta`
+- Canonical task record: `.agentplane/tasks/202605132058-36REEV/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-13T20:59:57.460Z
+- Branch: task/202605132058-36REEV/split-pr-meta
+- Head: 403e10e9a434
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
